feat: Enabling custom endpoints
An Express application created with the existing http.Server. This express server is now enabling easly to extend new endpoints to the HTTP server along with the Playwright server. For instance there's only a test endpoint added for POC reaosons. Client code is also extended to test if it's working